+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## 08 (October 2021)
+* See dashboards in a workbook
+* Add shapes property
+* Add custom sql
+* Drop python 2, add up through 3.9
+
+
 ## 07 (26 May 2021)
 * Fix bug in xfile that overwrote the namespace name when saving a document